+++ clickhouse-http-client/src/main/java/com/clickhouse/client/http/ClickHouseHttpConnection.java
@@ -10,11 +10,15 @@ import java.net.URLEncoder;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
+
+import org.ietf.jgss.GSSException;
+
 import java.util.Map.Entry;
 
 import com.clickhouse.client.ClickHouseClient;
@@ -25,6 +29,7 @@ import com.clickhouse.client.ClickHouseRequest;
 import com.clickhouse.client.ClickHouseRequestManager;
 import com.clickhouse.client.config.ClickHouseClientOption;
 import com.clickhouse.client.config.ClickHouseProxyType;
+import com.clickhouse.client.gss.GssAuthorizationContext;
 import com.clickhouse.client.http.config.ClickHouseHttpOption;
 import com.clickhouse.config.ClickHouseOption;
 import com.clickhouse.data.ClickHouseByteUtils;
@@ -46,6 +51,8 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
             .getBytes(StandardCharsets.US_ASCII);
     private static final byte[] HEADER_BINARY_ENCODING = "content-transfer-encoding: binary\r\n\r\n"
             .getBytes(StandardCharsets.US_ASCII);
+
+    private static final String HEADER_AUTHORIZATION = "authorization";
 
     private static final byte[] ERROR_MSG_PREFIX = "ode: ".getBytes(StandardCharsets.US_ASCII);
 
@@ -200,7 +207,7 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
             if (value == null) {
                 continue;
             }
-            if ("authorization".equals(name)) {
+            if (HEADER_AUTHORIZATION.equals(name)) {
                 hasAuthorizationHeader = true;
             }
             map.put(name, value);
@@ -212,16 +219,18 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
         }
         map.put("user-agent", !ClickHouseChecker.isNullOrEmpty(userAgent) ? userAgent : config.getClientName());
 
-        ClickHouseCredentials credentials = server.getCredentials(config);
-        if (credentials.useAccessToken()) {
-            // TODO check if auth-scheme is available and supported
-            map.put("authorization", credentials.getAccessToken());
-        } else if (!hasAuthorizationHeader) {
-            map.put("x-clickhouse-user", credentials.getUserName());
-            if (config.isSsl() && !ClickHouseChecker.isNullOrEmpty(config.getSslCert())) {
-                map.put("x-clickhouse-ssl-certificate-auth", "on");
-            } else if (!ClickHouseChecker.isNullOrEmpty(credentials.getPassword())) {
-                map.put("x-clickhouse-key", credentials.getPassword());
+        ClickHouseCredentials credentials = getCredentials(config, server);
+        if (!credentials.isGssEnabled()) {
+            if (credentials.useAccessToken()) {
+                // TODO check if auth-scheme is available and supported
+                map.put(HEADER_AUTHORIZATION, credentials.getAccessToken());
+            } else if (!hasAuthorizationHeader) {
+                map.put("x-clickhouse-user", credentials.getUserName());
+                if (config.isSsl() && !ClickHouseChecker.isNullOrEmpty(config.getSslCert())) {
+                    map.put("x-clickhouse-ssl-certificate-auth", "on");
+                } else if (!ClickHouseChecker.isNullOrEmpty(credentials.getPassword())) {
+                    map.put("x-clickhouse-key", credentials.getPassword());
+                }
             }
         }
 
@@ -239,6 +248,10 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
             map.put("content-encoding", config.getRequestCompressAlgorithm().encoding());
         }
         return map;
+    }
+
+    private static ClickHouseCredentials getCredentials(ClickHouseConfig config, ClickHouseNode server) {
+        return server.getCredentials(config);
     }
 
     protected static Proxy getProxy(ClickHouseConfig config) {
@@ -350,10 +363,11 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
     protected final ClickHouseRequestManager rm;
 
     protected final ClickHouseConfig config;
-    protected final Map<String, String> defaultHeaders;
     protected final String url;
+    protected final Map<String, String> defaultHeaders;
+    protected final GssAuthorizationContext gssAuthContext;
 
-    protected ClickHouseHttpConnection(ClickHouseNode server, ClickHouseRequest<?> request) {
+    protected ClickHouseHttpConnection(ClickHouseNode server, ClickHouseRequest<?> request, GssAuthorizationContext gssAuthContext) {
         if (server == null || request == null) {
             throw new IllegalArgumentException("Non-null server and request are required");
         }
@@ -363,9 +377,14 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
 
         ClickHouseConfig c = request.getConfig();
         this.config = c;
-        this.defaultHeaders = Collections.unmodifiableMap(createDefaultHeaders(c, server, getUserAgent()));
         this.url = buildUrl(server.getBaseUri(), request);
+        this.defaultHeaders = Collections.unmodifiableMap(createDefaultHeaders(c, server, getUserAgent()));
         log.debug("url [%s]", this.url);
+        this.gssAuthContext = gssAuthContext;
+    }
+
+    protected GssAuthorizationContext getGssAuthorizationContext() {
+        return gssAuthContext;
     }
 
     protected void closeQuietly() {
@@ -416,13 +435,13 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
      */
     protected Map<String, String> mergeHeaders(Map<String, String> requestHeaders) {
         if (requestHeaders == null || requestHeaders.isEmpty()) {
-            return defaultHeaders;
+            return setGssAuthHeader(new HashMap<>(defaultHeaders), config, server);
         } else if (isReusable()) {
-            return requestHeaders;
+            return setGssAuthHeader(requestHeaders, config, server);
         }
 
         Map<String, String> merged = new LinkedHashMap<>();
-        merged.putAll(defaultHeaders);
+        merged.putAll(requestHeaders);
         for (Entry<String, String> header : requestHeaders.entrySet()) {
             String name = header.getKey().toLowerCase(Locale.ROOT);
             String value = header.getValue();
@@ -432,7 +451,31 @@ public abstract class ClickHouseHttpConnection implements AutoCloseable {
                 merged.put(name, value);
             }
         }
+        setGssAuthHeader(merged, config, server);
         return merged;
+    }
+
+    private Map<String, String> setGssAuthHeader(Map<String, String> headers, ClickHouseConfig config,
+            ClickHouseNode server) {
+        if (headers.containsKey(HEADER_AUTHORIZATION)) {
+            log.warn("Authorization header is present. Skipping");
+            return headers;
+        }
+        ClickHouseCredentials credentials = getCredentials(config, server);
+        if (credentials.isGssEnabled()) {
+            if (gssAuthContext == null) {
+                throw new IllegalStateException("GssAuthorizer not initialized");
+            }
+            String userName = credentials.getUserName();
+            try {
+                headers.put(HEADER_AUTHORIZATION, "Negotiate "
+                        + gssAuthContext.getAuthToken(userName, config.getKerberosServerName(), server.getHost()));
+            } catch (GSSException e) {
+                throw new RuntimeException("Can not generate GSS token for user " + userName + " host "
+                        + server.getHost() + " with name " + config.getKerberosServerName(), e);
+            }
+        }
+        return headers;
     }
 
     /**