+++ clickhouse-http-client/src/main/java11/com/clickhouse/client/http/ClickHouseHttpConnectionFactory.java
@@ -5,6 +5,7 @@ import java.util.concurrent.ExecutorService;
 
 import com.clickhouse.client.ClickHouseNode;
 import com.clickhouse.client.ClickHouseRequest;
+import com.clickhouse.client.gss.GssAuthorizationContext;
 import com.clickhouse.client.http.config.ClickHouseHttpOption;
 import com.clickhouse.client.http.config.HttpConnectionProvider;
 import com.clickhouse.logging.Logger;
@@ -14,20 +15,20 @@ public final class ClickHouseHttpConnectionFactory {
     private static final Logger log = LoggerFactory.getLogger(ClickHouseHttpConnectionFactory.class);
 
     public static ClickHouseHttpConnection createConnection(ClickHouseNode server, ClickHouseRequest<?> request,
-            ExecutorService executor) throws IOException {
+            ExecutorService executor, GssAuthorizationContext gssAuthContext) throws IOException {
         HttpConnectionProvider provider = request.getConfig().getOption(ClickHouseHttpOption.CONNECTION_PROVIDER,
                 HttpConnectionProvider.class);
         if (provider == HttpConnectionProvider.APACHE_HTTP_CLIENT) {
             try {
-                return new ApacheHttpConnectionImpl(server, request, executor);
+                return new ApacheHttpConnectionImpl(server, request, executor, gssAuthContext);
             } catch (ExceptionInInitializerError | NoClassDefFoundError t) {
                 log.warn("Error when creating %s, fall back to HTTP_URL_CONNECTION", provider, t);
             }
         } else if (provider == HttpConnectionProvider.HTTP_CLIENT) {
-            return new HttpClientConnectionImpl(server, request, executor);
+            return new HttpClientConnectionImpl(server, request, executor, gssAuthContext);
         }
 
-        return new HttpUrlConnectionImpl(server, request, executor);
+        return new HttpUrlConnectionImpl(server, request, executor, gssAuthContext);
     }
 
     private ClickHouseHttpConnectionFactory() {